backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
motor
passlib[bcrypt]
python-jose[cryptography]
